testing/web-platform/mozilla/tests/indic-detection/tam.html

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
ªêšõ£Œ (Mars) ÅKò‚°´‹ðˆF™ àœ÷ å¼ «è£œ Ý°‹. Þ¶ ÅKòQL¼‰¶ ï£¡è£õ¶ «è£÷£è àœ÷¶.Þ‚°´‹ðˆF™ Iè„CPò «è£÷£ù ¹îÂ‚° Ü´ˆîî£è Þó‡ì£õ¶ CPò «è£÷£è ªêšõ£Œ Þ¼‚Aø¶. «ñù£†®ù˜ Þ‚«è£À‚° «ð£˜‚èì¾O¡ ªðò¬ó„ Å†®»œ÷ù˜. Þî¡ «ñŸðóŠH™ è£íŠð´‹ Þ¼‹¹ Ý‚¬ê´ Þ‚«è£¬÷„ ªê‰Gøñ£è‚ è£†´Aø¶.[12] Þîù£«ô«ò ÞîŸ°„ ªêšõ£Œ â¡ø ªðò˜ ãŸð†ì¶. å¼ ¹Mê£˜ «è£÷£ù Þî¡ «ñŸðóŠ¹ ê‰FóQ™ àœ÷¶«ð£™ A‡í‚ °Nè¬÷»‹, ¹MJ™ àœ÷¶ «ð£¡ø âKñ¬ôèœ, ðœ÷ˆî£‚°èœ, ð£¬ôõùƒèœ, ðQÍ®ò ¶¼õŠ ð°Fè¬÷»‹ ªè£‡ì¶. ªêšõ£J¡ ²öŸC‚è£ôº‹, ð¼õ ñ£ŸøƒèÀ‹ ¹M‚° àœ÷¬îŠ «ð£¡ø¬õ«ò. ÅKò ñ‡ìôˆ¶œ Iè àòóñ£ù åL‹ð² ñ¬ô»‹, IèŠªðKò ªêƒ°ˆ¶Š ðœ÷ˆî£‚°èÀœ å¡ø£ù ñKù˜ ðœ÷ˆî£‚°‹ ªêšõ£J«ô«ò àœ÷ù.
<script>
test(function() {
  assert_equals(document.characterSet, "windows-1252");
},"Should fall back to windows-1252");
</script>
The text content above is a converted extract from the start of the Wikipedia article <a href="https://ta.wikipedia.org/w/index.php?title=%E0%AE%9A%E0%AF%86%E0%AE%B5%E0%AF%8D%E0%AE%B5%E0%AE%BE%E0%AE%AF%E0%AF%8D_(%E0%AE%95%E0%AF%8B%E0%AE%B3%E0%AF%8D)&oldid=3129711">&#2970;&#3014;&#2997;&#3021;&#2997;&#3006;&#2991;&#3021; (&#2965;&#3019;&#2995;&#3021;)</a> and
is licensed under the <a href="https://creativecommons.org/licenses/by-sa/3.0/legalcode">Creative Commons Attribution-ShareAlike 3.0 Unported</a> license.
